forgor those

+++ src/m_menu.c
@@ -10832,7 +10832,7 @@ static void M_DrawSetupMultiPlayerMenu(void)
 				INT32 x = ((s % SKINGRIDWIDTH) * 18) + ((BASEVIDWIDTH / 2) - (18 * SKINGRIDWIDTH) - 8) + 100 + SKINXSHIFT; //BASEVIDWIDTH / 2 - ((icons + 1) * 24) - 4;
 				INT32 y = ((s / SKINGRIDWIDTH) * 18) + ((BASEVIDHEIGHT / 2) - (18 * (SKINGRIDWIDTH/2))); //BASEVIDWIDTH / 2 - ((icons + 1) * 24) - 4;
 				INT32 calcs = s + (setupm_skinypos * SKINGRIDWIDTH);
-				INT16 skinn;
+				INT32 skinn;
 				patch_t *face;
 				UINT8 *cmap;
